move from graphql mutation to rest post #25833
