+++ src/schema/inputs/index.ts
@@ -2,4 +2,3 @@ export * from './position-attribute';
 export * from './user-profile';
 export * from './user.input';
 export * from './layout.input';
-export * from './grid-settings';
